HTML and CSS Validation
HTML and CSS code is now compliant with the appropriate standards.

Styles.css changed from flex-direction: row wrap; to flex-direction: row;

I believe this was an error as the flex-wrap is already declared above it, however, it was non-compliant with it specified in direction.

Some changes from main to the crawler stuff also being updated here.

Database re-upload.